requirements.txt
comet-ml
easydict
numpy
python-dotenv
pytorch-lightning
PyYAML